first version of FFTAnalyzer + main loop update

+++ interface/WFClass.h
@@ -47,34 +47,34 @@ public:
     ~WFClass() {};
 
     //---getters---
-    inline vector<float>* GetSamples() {return &samples_;};
-    inline float          GetBaseline() {return baseline_;}
-    float                 GetAmpMax(int min=-1, int max=-1);
-    WFFitResults          GetInterpolatedAmpMax(int min=-1, int max=-1, int nFitSamples=7);
-    pair<float, float>    GetTime(string method, vector<float>& params); 
-    pair<float, float>    GetTimeCF(float frac, int nFitSamples=5, int min=-1, int max=-1);
-    pair<float, float>    GetTimeLE(float thr, int nmFitSamples=1, int npFitSamples=3, int min=-1, int max=-1);
-    float                 GetIntegral(int min=-1, int max=-1);
-    float                 GetModIntegral(int min=-1, int max=-1);
-    virtual float         GetSignalIntegral(int riseWin, int fallWin);
+    inline vector<double>* GetSamples() {return &samples_;};
+    inline float           GetBaseline() {return baseline_;}
+    float                  GetAmpMax(int min=-1, int max=-1);
+    WFFitResults           GetInterpolatedAmpMax(int min=-1, int max=-1, int nFitSamples=7);
+    pair<float, float>     GetTime(string method, vector<float>& params); 
+    pair<float, float>     GetTimeCF(float frac, int nFitSamples=5, int min=-1, int max=-1);
+    pair<float, float>     GetTimeLE(float thr, int nmFitSamples=1, int npFitSamples=3, int min=-1, int max=-1);
+    float                  GetIntegral(int min=-1, int max=-1);
+    float                  GetModIntegral(int min=-1, int max=-1);
+    virtual float          GetSignalIntegral(int riseWin, int fallWin);
     //---setters---
-    void                  SetSignalWindow(int min, int max);
-    void                  SetBaselineWindow(int min, int max);
-    void                  SetTemplate(TH1* templateWF=NULL);
+    void                   SetSignalWindow(int min, int max);
+    void                   SetBaselineWindow(int min, int max);
+    void                   SetTemplate(TH1* templateWF=NULL);
     //---utils---
-    void                  Reset();
-    void                  AddSample(float sample) {samples_.push_back(polarity_*sample);};
-    WFBaseline            SubtractBaseline(int min=-1, int max=-1);
-    WFFitResults          TemplateFit(float offset=0., int lW=0, int hW=0);
-    void                  EmulatedWF(WFClass& wf, float rms, float amplitude, float time);
-    void                  FFT(WFClass& wf, float tau, int cut);
-    void                  Print();
+    void                   Reset();
+    void                   AddSample(float sample) {samples_.push_back(polarity_*sample);};
+    WFBaseline             SubtractBaseline(int min=-1, int max=-1);
+    WFFitResults           TemplateFit(float offset=0., int lW=0, int hW=0);
+    void                   EmulatedWF(WFClass& wf, float rms, float amplitude, float time);
+    void                   FFT(WFClass& wf, float tau, int cut);
+    void                   Print();
     //---operators---
-    WFClass&              operator=(const WFClass& origin);
-    WFClass               operator-(const WFClass& sub);
-    WFClass               operator+(const WFClass& add);
-    WFClass&              operator-=(const WFClass& sub);
-    WFClass&              operator+=(const WFClass& add);
+    WFClass&               operator=(const WFClass& origin);
+    WFClass                operator-(const WFClass& sub);
+    WFClass                operator+(const WFClass& add);
+    WFClass&               operator-=(const WFClass& sub);
+    WFClass&               operator+=(const WFClass& add);
 protected:
     //---utils---
     float                 BaselineRMS();
@@ -82,31 +82,31 @@ protected:
     double                TemplateChi2(const double* par=NULL);
     
 protected:
-    vector<float> samples_;
-    float         tUnit_;
-    int           polarity_;
-    int           sWinMin_;
-    int           sWinMax_;
-    int           bWinMin_;
-    int           bWinMax_;
-    int           maxSample_;
-    float         fitAmpMax_;
-    float         fitTimeMax_;
-    float         fitChi2Max_;
-    float         baseline_;
-    float         bRMS_;
-    int           cfSample_;
-    float         cfFrac_;
-    float         cfTime_;
-    int           leSample_;
-    float         leThr_;
-    float         leTime_;
-    float         chi2cf_;
-    float         chi2le_;
-    int           fWinMin_;
-    int           fWinMax_;
-    float         tempFitTime_;
-    float         tempFitAmp_;
+    vector<double> samples_;
+    float          tUnit_;
+    int            polarity_;
+    int            sWinMin_;
+    int            sWinMax_;
+    int            bWinMin_;
+    int            bWinMax_;
+    int            maxSample_;
+    float          fitAmpMax_;
+    float          fitTimeMax_;
+    float          fitChi2Max_;
+    float          baseline_;
+    float          bRMS_;
+    int            cfSample_;
+    float          cfFrac_;
+    float          cfTime_;
+    int            leSample_;
+    float          leThr_;
+    float          leTime_;
+    float          chi2cf_;
+    float          chi2le_;
+    int            fWinMin_;
+    int            fWinMax_;
+    float          tempFitTime_;
+    float          tempFitAmp_;
     ROOT::Math::Interpolator* interpolator_;
 };
